[TC] Regenerated demo for tc_tc_capture_mode_sam_e54_xpro
MH3-58606

+++ apps/tc/tc_capture_mode/firmware/src/config/sam_e54_xpro/peripheral/tc/plib_tc1.c
@@ -85,7 +85,7 @@ void TC1_CaptureInitialize( void )
                                   | TC_CTRLA_CAPTEN0_Msk | TC_CTRLA_CAPTEN1_Msk  ;
 
 
-    TC1_REGS->COUNT16.TC_EVCTRL = (uint8_t)(TC_EVCTRL_EVACT_PPW | TC_EVCTRL_TCEI_Msk);
+    TC1_REGS->COUNT16.TC_EVCTRL = (uint16_t)(TC_EVCTRL_EVACT_PPW | TC_EVCTRL_TCEI_Msk);
 
     /* Clear all interrupt flags */
     TC1_REGS->COUNT16.TC_INTFLAG = (uint8_t)TC_INTFLAG_Msk;
